fix: flatc-drift-gates (#1240)
* ci: marmot + nmp.kernel flatc drift gates (#1240, #1288)

Extends the codegen-drift workflow so every checked-in FlatBuffers binding
is byte-diffed against a fresh pinned-flatc run, closing two silent-drift
gaps the existing nmp_update.fbs-only gates left open.

PART A (#1240) — marmot (NMMS):
- The checked-in Rust marmot bindings were raw flatc output, NOT the
  flatc + `rustfmt --edition 2021` form the proven nmp_update.fbs Rust gate
  asserts. Regenerated marmot_{snapshot,messages}_generated.rs byte-identical
  to canonical flatc 25.12.19 + rustfmt (nmp-marmot still builds).
- New ci/check-marmot-flatc-drift.sh with rust|swift|kotlin modes:
  - rust:   flatc --rust 25.12.19 + rustfmt diff
  - swift:  flatc --swift 25.12.19, renaming the snake_case flatc output to
            the checked-in PascalCase Marmot{Snapshot,Messages}.generated.swift
            before diffing
  - kotlin: flatc --kotlin 25.2.10, diff -r android/.../nmp/marmot
- Wired as three jobs in codegen-drift.yml reusing the existing pinned-flatc
  install steps; pins stay consistent with check-flatbuffers-version-pins.sh.

PART B (#1288) — Kotlin nmp.kernel:
- check-kotlin-flatc-drift.sh now also covers android/.../nmp/kernel. The
  nmp.kernel root schema list is DERIVED by grepping `namespace nmp.kernel`
  across crates/nmp-core/schema/*.fbs (never hard-coded), mirroring the TS
  gate's KRPR block. Each checked-in kernel .kt is byte-diffed against fresh
  flatc 25.2.10 output (the Android shell wires a subset of projections, so
  extra not-yet-wired emitted tables are not treated as drift).
- This caught genuine drift: the checked-in kernel bindings were stale —
  RelayDiagnosticsRow was missing the ADR-0051 `info:RelayDiagnosticsInfo`
  field and RelayDiagnosticsInfo.kt was never checked in. Regenerated the 8
  drifted kernel bindings byte-identical to canonical 25.2.10 and added the
  missing RelayDiagnosticsInfo.kt. All changes are additive/format-only; the
  hand-written TypedRelayDiagnosticsDecoder still compiles.

Updated the codegen-drift.yml header + regenerate instructions.

Fixes #1240, Fixes #1288


* fix(android): adapt action stage timestamp decoder to ulong

---------

+++ android/app/src/main/java/org/nmp/android/TypedActionStagesDecoder.kt
@@ -66,7 +66,7 @@ object TypedActionStagesDecoder {
                     stages.add(
                         ActionStageEntry(
                             stage = stage.stage ?: "",
-                            atMs = stage.atMs,
+                            atMs = stage.atMs.toLong(),
                             reason = if (stage.hasReason) stage.reason else null,
                         )
                     )
